Don't record/match headers; Update tables matchers
