+++ sdk/tables/azure-data-tables/tests/test_table_aad_async.py
@@ -9,7 +9,7 @@ import pytest
 from datetime import datetime
 import sys
 
-from devtools_testutils import AzureRecordedTestCase, set_bodiless_matcher
+from devtools_testutils import AzureRecordedTestCase, set_custom_default_matcher
 from devtools_testutils.aio import recorded_by_proxy_async
 
 from azure.data.tables.aio import TableServiceClient, TableClient
@@ -176,7 +176,10 @@ class TestTableAADAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_aad_batch_all_operations_together(self, tables_storage_account_name):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         await self._set_up(tables_storage_account_name, self.get_token_credential())
         try:
