+++ sdk/tables/azure-data-tables/tests/test_table_batch.py
@@ -12,7 +12,7 @@ from datetime import datetime, timedelta
 import os
 import sys
 
-from devtools_testutils import AzureRecordedTestCase, recorded_by_proxy, set_bodiless_matcher
+from devtools_testutils import AzureRecordedTestCase, recorded_by_proxy, set_custom_default_matcher
 
 from azure.core import MatchConditions
 from azure.core.pipeline.policies import HTTPPolicy
@@ -48,7 +48,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_single_insert(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -82,7 +85,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_single_update(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -120,7 +126,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_update(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -160,7 +169,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_merge(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -202,7 +214,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_update_if_match(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -233,7 +248,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_update_if_doesnt_match(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -264,7 +282,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_single_op_if_doesnt_match(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -311,7 +332,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_insert_replace(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -344,7 +368,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_insert_merge(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -377,7 +404,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_delete(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -412,7 +442,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_inserts(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -450,7 +483,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_all_operations_together(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -523,7 +559,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_reuse(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -565,7 +604,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_same_row_operations_fail(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -596,7 +638,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_different_partition_operations_fail(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -624,7 +669,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_too_many_ops(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -650,7 +698,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_different_partition_keys(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -670,7 +721,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_new_non_existent_table(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -691,7 +745,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_new_invalid_key(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         invalid_key = tables_primary_storage_account_key.named_key.key[0:-6] + "==" # cut off a bit from the end to invalidate
         tables_primary_storage_account_key = AzureNamedKeyCredential(tables_storage_account_name, invalid_key)
@@ -711,7 +768,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_new_delete_nonexistent_entity(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -730,7 +790,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_delete_batch_with_bad_kwarg(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -761,7 +824,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @pytest.mark.live_test_only
     @tables_decorator
     def test_batch_sas_auth(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -813,7 +879,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @pytest.mark.live_test_only  # Request bodies are very large
     @tables_decorator
     def test_batch_request_too_large(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -840,7 +909,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_with_mode(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -887,7 +959,10 @@ class TestTableBatch(AzureRecordedTestCase, TableTestCase):
     @tables_decorator
     @recorded_by_proxy
     def test_batch_with_specialchar_partitionkey(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
