+++ sdk/tables/azure-data-tables/tests/test_table_batch_async.py
@@ -12,7 +12,7 @@ from datetime import datetime, timedelta
 import os
 import sys
 
-from devtools_testutils import AzureRecordedTestCase, set_bodiless_matcher
+from devtools_testutils import AzureRecordedTestCase, set_custom_default_matcher
 from devtools_testutils.aio import recorded_by_proxy_async
 
 from azure.core import MatchConditions
@@ -44,7 +44,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_single_insert(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -77,7 +80,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_single_update(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -115,7 +121,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_update(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -152,7 +161,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_merge(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -193,7 +205,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_update_if_match(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -223,7 +238,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_update_if_doesnt_match(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -254,7 +272,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_insert_replace(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -286,7 +307,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_insert_merge(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -318,7 +342,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_delete(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -353,7 +380,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_inserts(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -393,7 +423,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_all_operations_together(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -468,7 +501,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_same_row_operations_fail(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -496,7 +532,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_different_partition_operations_fail(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -524,7 +563,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_too_many_ops(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -549,7 +591,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_new_non_existent_table(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -569,7 +614,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_new_invalid_key(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         invalid_key = tables_primary_storage_account_key.named_key.key[0:-6] + "==" # cut off a bit from the end to invalidate
         tables_primary_storage_account_key = AzureNamedKeyCredential(tables_storage_account_name, invalid_key)
@@ -587,7 +635,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_new_delete_nonexistent_entity(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -604,7 +655,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @pytest.mark.live_test_only
     @tables_decorator_async
     async def test_batch_sas_auth(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -654,7 +708,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @pytest.mark.live_test_only  # Request bodies are very large
     @tables_decorator_async
     async def test_batch_request_too_large(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -681,7 +738,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_delete_batch_with_bad_kwarg(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -712,7 +772,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_with_mode(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -759,7 +822,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_batch_with_specialchar_partitionkey(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
@@ -820,7 +886,10 @@ class TestTableBatchAsync(AzureRecordedTestCase, AsyncTableTestCase):
     @tables_decorator_async
     @recorded_by_proxy_async
     async def test_async_batch_inserts(self, tables_storage_account_name, tables_primary_storage_account_key):
-        set_bodiless_matcher()
+        # this can be reverted to set_bodiless_matcher() after tests are re-recorded and don't contain these headers
+        set_custom_default_matcher(
+            compare_bodies=False, excluded_headers="Authorization,Content-Length,x-ms-client-request-id,x-ms-request-id"
+        )
 
         # Arrange
         await self._set_up(tables_storage_account_name, tables_primary_storage_account_key)
