+++ tools/azure-sdk-tools/devtools_testutils/proxy_startup.py
@@ -18,6 +18,7 @@ import subprocess
 
 from .config import PROXY_URL
 from .helpers import is_live_and_not_recording
+from .sanitizers import add_remove_header_sanitizer, set_custom_default_matcher
 
 if TYPE_CHECKING:
     from typing import Optional
@@ -178,6 +179,11 @@ def start_test_proxy():
 
         # Wait for the proxy server to become available
         check_proxy_availability()
+        # remove headers from recordings if we don't need them, and ignore them if present
+        # Authorization, for example, can contain sensitive info and can cause matching failures during challenge auth
+        headers_to_ignore = "Authorization, x-ms-client-request-id, x-ms-request-id"
+        add_remove_header_sanitizer(headers=headers_to_ignore)
+        set_custom_default_matcher(excluded_headers=headers_to_ignore)
 
 
 def stop_test_proxy():
