+++ tools/azure-sdk-tools/devtools_testutils/proxy_testcase.py
@@ -3,7 +3,6 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-import os
 import logging
 import requests
 import six
@@ -18,7 +17,6 @@ except:
     import urlparse as url_parse
 
 import pytest
-import subprocess
 
 from azure.core.exceptions import HttpResponseError, ResourceNotFoundError
 from azure.core.pipeline.policies import ContentDecodePolicy
@@ -27,8 +25,9 @@ from azure.core.pipeline.transport import RequestsTransport
 
 # the trimming function to clean up incoming arguments to the test function we are wrapping
 from azure_devtools.scenario_tests.utilities import trim_kwargs_from_test_function
-from .helpers import is_live, is_live_and_not_recording
 from .config import PROXY_URL
+from .helpers import get_test_id, is_live, is_live_and_not_recording, set_recording_id
+from .sanitizers import add_remove_header_sanitizer, set_custom_default_matcher
 
 if TYPE_CHECKING:
     from typing import Tuple
@@ -42,42 +41,6 @@ RECORDING_START_URL = "{}/record/start".format(PROXY_URL)
 RECORDING_STOP_URL = "{}/record/stop".format(PROXY_URL)
 PLAYBACK_START_URL = "{}/playback/start".format(PROXY_URL)
 PLAYBACK_STOP_URL = "{}/playback/stop".format(PROXY_URL)
-
-# we store recording IDs in a module-level variable so that sanitizers can access them
-# we map test IDs to recording IDs, rather than storing only the current test's recording ID, for parallelization
-this = sys.modules[__name__]
-this.recording_ids = {}
-
-
-def get_recording_id():
-    test_id = get_test_id()
-    return this.recording_ids.get(test_id)
-
-
-def get_test_id():
-    # type: () -> str
-    # pytest sets the current running test in an environment variable
-    # the path to the test can depend on the environment, so we can't assume this is the path from the repo root
-    setting_value = os.getenv("PYTEST_CURRENT_TEST")
-    path_to_test = os.path.normpath(setting_value.split(" ")[0])
-    full_path_to_test = os.path.abspath(path_to_test)
-
-    # walk up to the repo root by looking for "sdk" directory or root of file system
-    path_components = []
-    head, tail = os.path.split(full_path_to_test)
-    while tail != "sdk" and tail != "":
-        path_components.append(tail)
-        head, tail = os.path.split(head)
-
-    # reverse path_components to construct components of path from repo root: [sdk, ..., tests, {test}]
-    path_components.append("sdk")
-    path_components.reverse()
-    for idx, val in enumerate(path_components):
-        if val.startswith("test"):
-            path_components.insert(idx + 1, "recordings")
-            break
-
-    return os.sep.join(path_components).replace("::", "").replace("\\", "/")
 
 
 def start_record_or_playback(test_id):
@@ -121,7 +84,7 @@ def start_record_or_playback(test_id):
                 )
 
     # set recording ID in a module-level variable so that sanitizers can access it
-    this.recording_ids[test_id] = recording_id
+    set_recording_id(test_id, recording_id)
     return (recording_id, variables)
